refactor: using map for paragraph
