+++ package.json
@@ -8,8 +8,8 @@
         "build": "yarn lint && tsc && vite build",
         "preview": "vite preview",
         "prepare": "husky install",
-        "lint": "yarn eslint --fix --ext .js,.jsx,.ts,.tsx src/ && yarn prettier --write .",
-        "lint:auto": "yarn eslint --fix --no-color --ext .js,.jsx,.ts,.tsx src/ && yarn prettier --write --no-color ."
+        "lint": "yarn eslint --fix --ignore-pattern src/fonts/* src/**/* && yarn prettier --write .",
+        "lint:auto": "yarn eslint --fix --no-color --ext .js,.jsx,.ts,.tsx --ignore-pattern src/fonts/* src/**/* && yarn prettier --write --no-color ."
     },
     "dependencies": {
         "antd": "^4.21.6",
